+++ ui/src/app/modules/settings/settings.component.ts
@@ -1,6 +1,5 @@
 import type { NetworkAdapterAvailable, NetworkAdapterSelected } from '@/app/modules/settings/settings.interfaces'
 
-import { animate, style, transition, trigger } from '@angular/animations'
 import { TitleCasePipe } from '@angular/common'
 import { afterNextRender, ChangeDetectorRef, Component, createEnvironmentInjector, DestroyRef, ElementRef, EnvironmentInjector, inject, OnInit, runInInjectionContext, signal, viewChild } from '@angular/core'
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop'
@@ -27,6 +26,7 @@ import { RemoveIndividualAccessoriesComponent } from '@/app/modules/settings/rem
 import { ResetAllBridgesComponent } from '@/app/modules/settings/reset-all-bridges/reset-all-bridges.component'
 import { ResetIndividualBridgesComponent } from '@/app/modules/settings/reset-individual-bridges/reset-individual-bridges.component'
 import { SelectNetworkInterfacesComponent } from '@/app/modules/settings/select-network-interfaces/select-network-interfaces.component'
+import { settingsAnimations } from '@/app/modules/settings/settings.animations'
 import { SslSettingsModalComponent } from '@/app/modules/settings/ssl-settings-modal/ssl-settings-modal.component'
 import { WallpaperComponent } from '@/app/modules/settings/wallpaper/wallpaper.component'
 
@@ -42,43 +42,7 @@ import { WallpaperComponent } from '@/app/modules/settings/wallpaper/wallpaper.c
     TranslatePipe,
     SpinnerComponent,
   ],
-  animations: [
-    trigger('fadeInOut', [
-      transition(':enter', [
-        style({ opacity: 0 }),
-        animate('750ms', style({ opacity: 1 })),
-      ]),
-      transition(':leave', [
-        animate('750ms', style({ opacity: 0 })),
-      ]),
-    ]),
-    trigger('smoothHide', [
-      transition(':enter', [
-        style({
-          opacity: 0.01,
-          height: '0px',
-          marginBottom: 0,
-          transform: 'scale(0.97)',
-        }),
-        animate('450ms cubic-bezier(0.0, 0.0, 0.2, 1)', style({
-          opacity: 1,
-          height: '*',
-          marginBottom: '*',
-          transform: 'scale(1)',
-        })),
-      ]),
-      transition(':leave', [
-        animate('150ms cubic-bezier(0.4, 0.0, 1, 1)', style({
-          opacity: 0,
-          height: 0,
-          marginBottom: 0,
-          paddingTop: 0,
-          paddingBottom: 0,
-          transform: 'scale(0.98)',
-        })),
-      ]),
-    ]),
-  ],
+  animations: settingsAnimations,
 })
 export class SettingsComponent implements OnInit {
   // Injected dependencies